// src/models/character-statistics.ts
export interface CharacterStatistics {
  id: number;
  weight: number;
  gravity: number;
  walkSpeed: number;
  runSpeed: number;
  waveDashLengthRank: number;
  plaIntangibilityFrames: number;
  jumpSquat: number;
  canWallJump: boolean;
  notes: string;
  initialDash: number;
  dashFrames: number;
  waveDashLength: number;
}